requirements.txt
numpy==1.19.2
Pillow~=8.4.0
termcolor==1.1.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.62.3
yacs==0.1.8

opencv-python~=4.5.4.60